move animation extensions to package modifier
